Moved test and removed duplicated ones
