+++ conans/test/functional/toolchains/cmake/cmakedeps/test_cmakedeps_build_modules.py
@@ -268,92 +268,3 @@ class TestNoNamespaceTarget:
             assert '>> nonamespace libs: library::library' in t.out
             t.run_command('cmake --build . --config Release')  # Compiles and links.
 
-
-@pytest.mark.parametrize("check_components_exist", [False, True, None])
-def test_targets_declared_in_build_modules(check_components_exist):
-    """If a require is declaring the component targets in a build_module, CMakeDeps is
-       fine with it, not needed to locate it as a conan declared component"""
-
-    client = TestClient()
-    conanfile_hello = str(GenConanfile().with_name("hello").with_version("1.0")
-                          .with_exports_sources("*.cmake", "*.h"))
-    conanfile_hello += """
-    def package(self):
-        self.copy("*.h", dst="include")
-        self.copy("*.cmake", dst="cmake")
-
-    def package_info(self):
-        self.cpp_info.set_property("cmake_build_modules", ["cmake/my_modules.cmake"])
-    """
-    my_modules = textwrap.dedent("""
-    add_library(cool_component INTERFACE)
-    target_include_directories(cool_component INTERFACE ${CMAKE_CURRENT_LIST_DIR}/../include/)
-    add_library(hello::invented ALIAS cool_component)
-    """)
-    hello_h = "int cool_header_only=1;"
-    client.save({"conanfile.py": conanfile_hello,
-                 "my_modules.cmake": my_modules, "hello.h": hello_h})
-    client.run("create .")
-
-    conanfile = textwrap.dedent("""
-            from conans import ConanFile
-            from conan.tools.cmake import CMake, CMakeDeps
-
-            class HelloConan(ConanFile):
-                name = 'app'
-                version = '1.0'
-                exports_sources = "*.txt", "*.cpp"
-                generators = "CMakeToolchain"
-                requires = ("hello/1.0", )
-                settings = "os", "compiler", "arch", "build_type"
-
-                def generate(self):
-                    deps = CMakeDeps(self)
-                    {}
-                    deps.generate()
-
-                def build(self):
-                    cmake = CMake(self)
-                    cmake.configure()
-                    cmake.build()
-
-    """)
-    if check_components_exist is False:
-        conanfile = conanfile.format("deps.check_components_exist=False")
-    elif check_components_exist is True:
-        conanfile = conanfile.format("deps.check_components_exist=True")
-    else:
-        conanfile = conanfile.format("")
-
-    main_cpp = textwrap.dedent("""
-        #include <iostream>
-        #include "hello.h"
-
-        int main(){
-            std::cout << "cool header value: " << cool_header_only;
-            return 0;
-        }
-        """)
-
-    cmakelist = textwrap.dedent("""
-        set(CMAKE_CXX_COMPILER_WORKS 1)
-        set(CMAKE_CXX_ABI_COMPILED 1)
-        set(CMAKE_C_COMPILER_WORKS 1)
-        set(CMAKE_C_ABI_COMPILED 1)
-        cmake_minimum_required(VERSION 3.15)
-        project(project CXX)
-
-        find_package(hello COMPONENTS invented missing)
-        add_executable(myapp main.cpp)
-        target_link_libraries(myapp hello::invented)
-    """)
-    client.save({"conanfile.py": conanfile,
-                 "CMakeLists.txt": cmakelist, "main.cpp": main_cpp})
-    client.run("create .", assert_error=check_components_exist)
-    assert bool(check_components_exist) == ("Conan: Component 'missing' NOT found in package "
-                                            "'hello'" in client.out)
-
-    assert "Conan: Including build module" in client.out
-    assert "my_modules.cmake" in client.out
-    assert bool(check_components_exist) == ("Conan: Component 'invented' found in package 'hello'"
-                                            in client.out)
